refactor: remove unused file and references to unused file
