--- src/components/WarningPunctuations.json
@@ -1,7 +1,0 @@
-[
-  {
-    "keyword": "!{3,}",
-    "source": "https://www.theatlantic.com/technology/archive/2018/06/exclamation-point-inflation/563774/?utm_source=feed",
-    "message": "We get it, you're excited!!! Use that many exclamation points with your friends, not your coworkers or clients. This is not social media. --Julie Beck"
-  }
-]
